pyproject.toml
[tool.poetry]
name = "medplexity"
version = "0.1.5"
description = "medplexity helps with evaluation of LLMs for medical use-cases."
license = "MIT"
authors = ["MaksymPetyak <petyak.mi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.3.0"
datasets = "^2.14.5"
langchain = {version = "^0.0.306", optional = true}
google-search-results = {version = "^2.4.2", optional = true}
supabase = {version = "^2.0.3", optional = true}
python-dotenv = "^1.0.0"
openai = "^1.3.3"
pillow = "^10.1.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
black = "^23.3.0"
pre-commit = "^3.2.2"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
pytest-env = "^0.8.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocstrings-python = "^1.7.0"
markdown-include = "^0.8.1"
mkdocs-material = "^9.4.3"

[tool.poetry.extras]
langchain = ["langchain"]
google-search-results = ["google-search-results"]
supabase = ["supabase"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
